pyproject.toml
[project]
name = "automation-account-x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-ollama>=0.2.3",
    "orjson>=3.10.15",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.17",
    "tweepy>=4.15.0",
    "langgraph>=0.2.74",
    "langsmith>=0.3.8",
    "langchain>=0.3.19",
]
